split the client from the node
